update defaults metrics on change over time chart. Hide treatment filter if no data
