+++ src/interface/src/app/treatments/change-over-time-chart/change-over-time-chart.component.scss
@@ -63,3 +63,12 @@ canvas {
   padding: 10px;
   text-align: center;
 }
+
+.loading-overlay {
+  background-color: rgba(255, 255, 255, 0.38);
+  width: 100%;
+  height: 100%;
+  position: absolute;
+  z-index: 100;
+
+}
