+++ src/interface/src/app/treatments/change-over-time-chart/change-over-time-chart.component.ts
@@ -1,43 +1,15 @@
-import { Component, Input } from '@angular/core';
+import { Component } from '@angular/core';
 import { ChartConfiguration } from 'chart.js';
 import { AsyncPipe, NgIf } from '@angular/common';
 import { NgChartsModule } from 'ng2-charts';
-import { DirectImpactsStateService } from '../direct-impacts.state.service';
-import {
-  combineLatest,
-  distinctUntilChanged,
-  filter,
-  map,
-  switchMap,
-} from 'rxjs';
-import {
-  ImpactsMetricSlot,
-  Metric,
-  SLOT_COLORS,
-  SLOT_PALETTES,
-} from '../metrics';
-import { TreatmentsService } from '@services/treatments.service';
+import { map } from 'rxjs';
+import { ImpactsMetricSlot, SLOT_COLORS, SLOT_PALETTES } from '../metrics';
 import { UntilDestroy } from '@ngneat/until-destroy';
-import { TreatmentsState } from '../treatments.state';
-import deepEqual from 'fast-deep-equal';
-import { TreatmentPlan } from '@types';
 import { getBasicChartOptions, updateYAxisRange } from '@app/chart-helper';
-
-export interface ImpactsResultData {
-  year: number;
-  variable: string;
-  dividend: number;
-  divisor: number;
-  value: number;
-  delta: number;
-  relative_year: number;
-}
-
-export interface ChangeOverTimeChartItem {
-  variable: string;
-  year: number;
-  avg_value: number;
-}
+import {
+  ChangeOverTimeChartItem,
+  ChangeOverTimeChartService,
+} from './change-over-time-chart.service';
 
 @UntilDestroy()
 @Component({
@@ -48,17 +20,9 @@ export interface ChangeOverTimeChartItem {
   styleUrl: './change-over-time-chart.component.scss',
 })
 export class ChangeOverTimeChartComponent {
-  constructor(
-    private directImpactsStateService: DirectImpactsStateService,
-    private treatmentsState: TreatmentsState,
-    private treatmentsService: TreatmentsService
-  ) {}
-
-  @Input() metrics!: Record<ImpactsMetricSlot, Metric> | null;
+  constructor(public chartService: ChangeOverTimeChartService) {}
 
   baseOptions = getBasicChartOptions();
-
-  emptyData = false;
 
   readonly staticBarChartOptions: ChartConfiguration<'bar'>['options'] = {
     ...this.baseOptions,
@@ -69,22 +33,19 @@ export class ChangeOverTimeChartComponent {
         ...this.baseOptions?.scales?.['x'],
         title: {
           ...this.baseOptions?.scales?.['x']?.['title'],
-          ...{
-            padding: {
-              ...{ top: -20 },
-            },
-          },
+          padding: { top: -20 },
         },
       },
     },
   };
 
-  chartConfiguration(
+  barChartData$ = this.chartService.barChartData$.pipe(
+    map((data) => (data ? this.chartConfiguration(data) : undefined))
+  );
+
+  private chartConfiguration(
     data: Record<ImpactsMetricSlot, ChangeOverTimeChartItem[]>
-  ) {
-    if (!data || this.emptyData) {
-      return undefined;
-    }
+  ): ChartConfiguration<'bar'>['data'] {
     const allValues = Object.values(data).flatMap((entries) =>
       entries.map((entry) => entry.avg_value)
     );
@@ -93,92 +54,26 @@ export class ChangeOverTimeChartComponent {
       labels: [0, 5, 10, 15, 20],
       datasets: [
         {
-          data: data['blue'].map((d: ChangeOverTimeChartItem) => d.avg_value),
+          data: data['blue'].map((d) => d.avg_value),
           backgroundColor: SLOT_COLORS['blue'],
           hoverBackgroundColor: SLOT_PALETTES['blue'][0],
         },
         {
-          data: data['purple'].map((d: ChangeOverTimeChartItem) => d.avg_value),
+          data: data['purple'].map((d) => d.avg_value),
           backgroundColor: SLOT_COLORS['purple'],
           hoverBackgroundColor: SLOT_PALETTES['purple'][0],
         },
         {
-          data: data['orange'].map((d: ChangeOverTimeChartItem) => d.avg_value),
+          data: data['orange'].map((d) => d.avg_value),
           backgroundColor: SLOT_COLORS['orange'],
           hoverBackgroundColor: SLOT_PALETTES['orange'][0],
         },
         {
-          data: data['green'].map((d: ChangeOverTimeChartItem) => d.avg_value),
+          data: data['green'].map((d) => d.avg_value),
           backgroundColor: SLOT_COLORS['green'],
           hoverBackgroundColor: SLOT_PALETTES['green'][0],
         },
       ],
-    } as ChartConfiguration<'bar'>['data'];
-  }
-
-  convertImpactResultToChartData(
-    resultData: ImpactsResultData[],
-    metrics: Record<ImpactsMetricSlot, Metric>
-  ): Record<ImpactsMetricSlot, ChangeOverTimeChartItem[]> {
-    const chartData = resultData.map((d) => {
-      return {
-        // map the returned data to data attributes for the map
-        year: d.relative_year,
-        avg_value: d.delta * 100,
-        variable: d.variable,
-      };
-    });
-    const dataBySlot = {
-      blue: chartData
-        .filter((item) => item.variable === metrics.blue.id)
-        .sort((a, b) => a.year - b.year),
-      green: chartData
-        .filter((item) => item.variable === metrics.green.id)
-        .sort((a, b) => a.year - b.year),
-      purple: chartData
-        .filter((item) => item.variable === metrics.purple.id)
-        .sort((a, b) => a.year - b.year),
-      orange: chartData
-        .filter((item) => item.variable === metrics.orange.id)
-        .sort((a, b) => a.year - b.year),
     };
-    return dataBySlot;
   }
-
-  isDataEmpty(givenData: any[]): boolean {
-    return givenData.every((e) => !e.value);
-  }
-
-  barChartData$ = combineLatest([
-    this.treatmentsState.treatmentPlan$.pipe(
-      filter((plan): plan is TreatmentPlan => !!plan)
-    ),
-    this.directImpactsStateService.reportMetrics$?.pipe(
-      distinctUntilChanged((prev, curr) => deepEqual(prev, curr))
-    ),
-    this.directImpactsStateService.selectedProjectArea$,
-    this.directImpactsStateService.filteredTreatmentTypes$,
-  ]).pipe(
-    switchMap(([plan, metrics, area, txTypes]) => {
-      const metricsArray = Object.values(metrics).map((m) => m.id);
-      let selectedArea = null;
-      if (area !== 'All') {
-        selectedArea = area.project_area_id;
-      }
-      this.emptyData = false;
-
-      return this.treatmentsService
-        .getTreatmentImpactCharts(plan.id, metricsArray, selectedArea, txTypes)
-        .pipe(
-          map((responseData) => ({ response: responseData, metrics: metrics }))
-        ); // we need the metrics for reviewing calculation, so passing it along here
-    }),
-    map((responseData) => {
-      const resultData = responseData.response as ImpactsResultData[];
-      this.emptyData = this.isDataEmpty(resultData);
-      return this.chartConfiguration(
-        this.convertImpactResultToChartData(resultData, responseData.metrics)
-      );
-    })
-  );
 }
