+++ src/interface/src/app/treatments/direct-impacts.state.service.spec.ts
@@ -1,5 +1,4 @@
 import { TestBed } from '@angular/core/testing';
-import { HttpClientTestingModule } from '@angular/common/http/testing';
 import { DirectImpactsStateService } from './direct-impacts.state.service';
 
 describe('DirectImpactsStateService', () => {
@@ -7,7 +6,6 @@ describe('DirectImpactsStateService', () => {
 
   beforeEach(() => {
     TestBed.configureTestingModule({
-      imports: [HttpClientTestingModule],
       providers: [DirectImpactsStateService],
     });
     service = TestBed.inject(DirectImpactsStateService);
