+++ src/interface/src/app/treatments/direct-impacts/direct-impacts.component.scss
@@ -206,6 +206,8 @@ $panel-height: 600px;
 }
 
 .project-area-selector {
+  min-height: 38px;
+
   ::ng-deep .mat-mdc-select-trigger {
     color: $color-black;
     font-size: 14px;
