+++ src/interface/src/app/treatments/direct-impacts/direct-impacts.component.ts
@@ -23,6 +23,7 @@ import { MetricFiltersComponent } from '@treatments/metric-filters/metric-filter
 import { ImpactsMetric } from '../metrics';
 
 import { DirectImpactsStateService } from '../direct-impacts.state.service';
+import { ChangeOverTimeChartService } from '@treatments/change-over-time-chart/change-over-time-chart.service';
 import { StandDataChartComponent } from '@treatments/stand-data-chart/stand-data-chart.component';
 import { Chart } from 'chart.js';
 import ChartDataLabels from 'chartjs-plugin-datalabels';
@@ -75,6 +76,7 @@ import { NavBarComponent } from '@app/standalone/nav-bar/nav-bar.component';
   ],
   providers: [
     DirectImpactsStateService,
+    ChangeOverTimeChartService,
     TreatmentsState,
     SelectedStandsState,
     TreatedStandsState,
@@ -106,9 +108,10 @@ export class DirectImpactsComponent implements OnInit, OnDestroy {
     private route: ActivatedRoute,
     private router: Router,
     private directImpactsStateService: DirectImpactsStateService,
+    private changeOverTimeChartService: ChangeOverTimeChartService,
     private fileSaverService: FileSaverService,
     private dialog: MatDialog,
-    private injector: Injector, // Angular's injector for passing shared services
+    private injector: Injector,
     private scenarioState: ScenarioState,
     private breadcrumbService: BreadcrumbService
   ) {
@@ -184,6 +187,8 @@ export class DirectImpactsComponent implements OnInit, OnDestroy {
       return `(Forested Stand)`;
     } else return `(Non-forested Stand)`;
   }
+
+  hasChartData$ = this.changeOverTimeChartService.hasChartData$;
 
   filterOptions$ = this.directImpactsStateService.reportMetrics$.pipe(
     map((metrics) => Object.values(metrics).map((metric) => metric.id))
